~ivy-posframe now applies to set of functions
Ivy-posframe now only applies to a set of functions through the
ivy-display-functions-alist. Some functions just aren't meant to be
posframed and are incredibly slow when using posframes, so I don't put
those in the list. Those functions in the list are functions that have a
negligible change in performance while providing a much better
experience via posframe.