~exported keymap settings to an elisp file
Keymap is a section of my config I edit quite a lot, and the org mode
format doesn't really fit with the way I edit it. Therefore I exported
it to an elisp file where I can quickly edit bits and pieces without a
lot of lag from explaining myself in org mode.