+brand new Emacs config
This one is a lot more closer to Emacs core, it doesn't use general
or the SPC leader as I'm used to but instead the default meta-key
chord bindings to work better. I don't use a lot of features in Emacs
that Doom used to provide in bindings, so I'm stripping them away and
slowly adding stuff I like.