~hyper -> super, caps -> escape
Remove the hyper key bindings, setting them purely to super.
After running pure Emacs with no leader bindings, I've found the use
of Ctrl+g/Escape really painful (as well as the use of Ctrl for
everything). I want to free the caps key (which was hyper) for use as escape.