(Emacs)+rg.el command to search in new frame
rg.el uses the compilation system to manage and display
results (similar to the *grep family of functions in Emacs).  I'd like
to be able to search in a separate frame (making that a 'searcher'
frame) instead of the current frame sometimes, so make a command that
does that for me.