(Emacs)+custom personalised ada-mode
The defautl ada-mode requires the installation of many different
tools, the compilation of several different systems and much more.  I
wasn't able to fix all the issues with my installation so I decided to
scrap it and make my own.  This mode just colourises the buffer so
far, but I hope to implement indentation shortly.