(xboxdrv)+configurations to make the xbox controller a working mouse
Xbox controllers are cool, and the ton of buttons allow for a lot of
customisation and ease of use.  Hence I created an xboxdrv
configuration which has the standard mouse capabilities as well as the
ability to access an on screen keyboard and some basic media
manipulation capabilities.  This is mostly for when I want to sit back
and just CONSOOM content rather than do actual work (hence trying to
remove the keyboard).