(Emacs)+haskell-multiedit
Just some code to make editing in the Haskell REPL a nicer
experience.  Allows me to write multi-line code in a separate buffer
with haskell-mode switched on, then with a simple keybinding paste and
evaluate the full code in the Haskell REPL.  The keybindings and
workflow are similar to org-src block editing (create a new buffer,
then paste the contents back into the original buffer).