(Emacs)+early-init native-compilation options
Set to 4 workers at the start to make compilation at least as fast as
possible: my laptop has 4 cores so it'll just slow it down until early
compilation is done, while my desktop can still keep going as it has 8
threads.  In config.org this is setup correctly so the rest of the
config (which is the much larger part) uses a system dependent number
of cores.