(Emacs)+org-link-frame-setup
Always open a new frame when opening link in org-mode: predictable,
safe behaviour which leaves it up to my window manager as to what to
do next.  Most of my opened links are files that I want to examine
along with the org file I opened them from: this allows me to do
exactly that.