(Emacs)~major change to +oreo/create-toggle-function
Now allows numeric arguments to be passed to the buffer creation
function.  This is especially useful for ~eshell~ which allows
creation of multiple instances using a numeric argument: very useful
for managing multiple buffers.  It's an optional argument but I err on
the side of caution so by default I'll set it to nil if I don't want
this feature.