(Emacs)~literate now actually checks if compilation is necessary
Just realised how easy a fix this is to the 2 second wait time for
Emacs to stop.  Checking if the org files are newer than the compiled
files, and the same for byte compilation, ensures I only compile when
necessary.

This actually makes Emacs kinda appealing for quick work: I kinda
cringed every time I launched Emacs without a server because I knew it
would take *so long* to stop it.  Now that isn't as big a concern!