(Doom)-Deleted Doom
I've superseded all the functionality I wanted in doom for quite some
time now.  I don't see any reason for using it, and it usually makes
it a bit annoying to find stuff in the project due to file name
conflicts.

As a final note, thank you very much Doom Emacs (hlissner in
particular) for introducing me to this amazing software.  The
community was really welcoming and I was very much eased into the
learning curve of Emacs, after coming from Vim.