(Emacs/elisp)~massively reworked bookmarking system
Now it can:
- cache results until bookmarks file is modified, so we don't need to
open the file again if we've already figured everything out.  This is
in-memory and not as a file

- a proper dispatching list which has pairs of form (PATTERNS . FUNC)
so the dispatching system is a bit more abstract and easier to add to

- use a URL property in org headings to get the bookmark URL rather
than having it as part of the body of the entry.  This means the body
is now free to house any content without affecting the overall
function of the system, such as notes or subtrees for related
bookmarks.