(Emacs/config)+display-fill-column-indicator-mode
This adds a strict column marker for the fill-column, which is nicer
than whitespace-mode's purple text marker (though I have both enabled
of course).  Turn it on for only certain modes, and set the fill
column to 80 by default but 70 in text mode buffers (because that's
what I'm used to).