(Scripts)~fix eselect
Big issue with emacsclient when launching alternate editor with eval
instructions: it considers the --eval Lisp code a file name, which it
opens as a buffer.  Instead, use -a "" which launches Emacs with
daemon on if no daemon is running.

Also, instead of using a SystemD service for Emacs, let's just manage
it via emacsclient.  Less dependence on it then.