Rework elfeed
All my feeds are stored in a separate org file.  I've written a parser
`elfeed-org` which parses the headings-as-links with the overarching
tags into a format for `elfeed-feeds`.

Why didn't I use the third-party elfeed-org package?  It uses an older
version of org which straight will clone entirely.  Don't want two
versions of org, and its a task just simple enough for me to do
myself.