Fix bug eshell-prompt
When on a remote directory (i.e. ssh'd onto a separate folder), user
input would be coloured based on the status colour of the last
command.  This colouring should only apply on the actual user-prompt
component - so this commit fixes it.

RCA: caused by run-on-colouring from the ep/user-prompt i.e. no
defaulting of the face.